Implemented paging
Implemented paging 3 on database, removed chatActivity.

getPendingIntent() is not available so far, will be fix in next push.
