+++ RecallRecall/app/src/main/java/com/recallrecall/app/db/MessageDao.kt
@@ -2,6 +2,8 @@ package com.recallrecall.app.db
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
+import androidx.paging.DataSource
+import androidx.paging.PagingSource
 import androidx.room.*
 import kotlinx.coroutines.flow.Flow
 
@@ -19,7 +21,7 @@ interface MessageDao {
     fun loadByDateAndName(startDate: String?, endDate: String?, name: String?): List<Message?>?
 
     @Query(value = "SELECT * FROM db_message WHERE name LIKE :name ORDER BY id DESC")
-    fun loadByName(name: String?): LiveData<List<Message?>?>
+    fun loadByName(name: String?): PagingSource<Int, Message>
 
     @Query(value = "SELECT * FROM db_message WHERE name LIKE :name AND recalled LIKE :recalled")
     fun loadByNameAndRecalled(name: String?, recalled: Boolean): List<Message?>?
